% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\title{Integrating with Trapezoid or Simpson's Rule}
\usage{
integrateIt(object)
}
\value{
A list with the input elements and the answer
 \item{object@x}{The first object input} 
 \item{object@y}{The second object input}
 \item{object@z}{The third object input}
 \item{Answer}{The value calculated using Trapezoid or Simpson's Rule}
}
\description{
Uses Trapezoid or Simpson's Rule to approximate the area under a set of points
}
\section{Slots}{

\describe{
\item{\code{object}}{An object of class Trapezoid or Integral}
}}

\examples{

integrateIt(newTrapezoid(1:4, c(3,6,12,13), "Trapezoid"))
}
\seealso{
newTrapezoid, newSimpson, printIntegral
}
\author{
Emily Garner<\email{emily.garner@wustl.edu}>
}
