% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newIntegral.R
\docType{class}
\name{Integral-class}
\alias{Integral-class}
\title{Create new Integral}
\value{
A list with the elements
 \item{x}{The first object input} 
 \item{y}{The second object input}
 \item{z}{The third object input}
}
\description{
Creates a new object of S4 class Integral
}
\section{Slots}{

\describe{
\item{\code{x}}{A numeric object indicating a set of points on the x axis increasing in value in equal increments}

\item{\code{y}}{A numeric object with the same dimensionality as \code{x} indicating a set of points on the y axis to be integrated over}

\item{\code{rule}}{A character object "Trapezoid" or "Simpson" indicating subclass}
}}

\examples{

myX <- seq(3,12,3) 
myY <- c(2, 5, 12, 8) 
newIntegral <- (myX, myY, "Trapezoid")
}
\seealso{
newTrapezoid, newSimpson
}
\author{
Emily Garner<\email{emily.garner@wustl.edu}>
}
